users/serializers.py
from rest_framework import serializers
from rest_framework_simplejwt.tokens import RefreshToken
from django.contrib.auth import authenticate
import logging

from users.models import User
from users.tasks import send_register_email

logger = logging.getLogger(__name__)


class UserRegisterSerializer(serializers.Serializer):
    email = serializers.EmailField(required=True)
    password = serializers.CharField(required=True)
    password_confirm = serializers.CharField(required=True)

    def validate(self, data):
        if data["password"] != data["password_confirm"]:
            raise serializers.ValidationError("Passwords do not match. Please insert correct passwords.")
        return data

    def save(self, **kwargs):
        user = User.objects._create_user(email=self.validated_data["email"],
        password=self.validated_data["password"],
        is_confirmed=False,)

        send_register_email.delay(user_id=user.id, email=user.email)

        return user


class UserLoginSerializer(serializers.Serializer):
    email = serializers.EmailField(required=True)
    password = serializers.CharField(write_only=True)

    def validate(self, data):
        email = data.get("email")
        password = data.get("password")

        user = authenticate(username=email, password=password)

        if not user:
            logger.warning(f"Failed login attempt with email: {email}")
            raise serializers.ValidationError("Invalid email or password.")

        if not user.is_confirmed:
            logger.warning(f"Unconfirmed email login attempt: {email}")
            raise serializers.ValidationError("Please confirm your email before logging in.")

        refresh = RefreshToken.for_user(user)
        return {
            "refresh": str(refresh),
            "access": str(refresh.access_token),
            "user": {
                "id": user.id,
                "email": user.email,
            },
        }